+++ src/main/java/cz/cvut/fit/miadp/mvcgame/model/gameobjects/familyA/MissileA.java
@@ -9,7 +9,7 @@ public class MissileA extends AbsMissile {
     
     IMovingStrategy movingStrategy;
 
-    public MissileA(Position initialPosition, double initAngle, int initVelocity, IMovingStrategy movingStrategy) {
+    public MissileA(Position initialPosition, double initAngle, double initVelocity, IMovingStrategy movingStrategy) {
         super(initialPosition, initAngle, initVelocity);
         this.movingStrategy = movingStrategy;
     }
